requirements.txt
Box2D==2.3.10
cElementTree==1.0.2-20050302
click==8.1.7
cma==4.0.0
commentjson==0.9.0
coverage==7.9.1
dataclass_dict_convert==1.7.3
datarobot==3.7.1
deap==1.4.2
elementtree==1.2.6-20050316
gym==0.26.2
h5py==3.12.1
hydra-core==1.3.2
ipython==8.12.3
lxml==6.0.0
matplotlib==3.10.3
mock==5.2.0
mpi4py==4.1.0
nnsight==0.3.6
numba==0.61.0
numexpr==2.10.1
omegaconf==2.1.2
omegaconf==2.3.0
ordered_set==4.1.0
pandas==2.3.0
Parser==0.1
pybullet==3.2.7
Pygments==2.19.2
pytest==8.3.5
python-dotenv==1.1.1
pythran==0.18.0
pytorch_lightning==2.4.0
runtests==0.0.28
scipy==1.16.0
seaborn==0.13.2
setuptools==72.1.0
stable_baselines==2.10.2
sympy==1.13.1
tensorflow==2.19.0
torch==2.5.1
tqdm==4.66.6
wandb==0.20.1
